sep into gpu_draw.py
