+++ Makefile
@@ -3,13 +3,13 @@
 ################################################################
 
 run: all
-	optirun blender -P blender.py ~/Documents/testspraypars.blend -p 0 0 940 1080
+	optirun blender -P blender.py ~/Documents/testlineview.blend -p 0 0 940 1080
 
 oldrun: all
-	optirun blender -P blender.py ~/Documents/testvoronoi.blend -p 0 0 940 1080
+	optirun blender -P blender.py ~/Documents/testspraypars.blend -p 0 0 940 1080
 
 baterun: all
-	optirun blender ~/Documents/Stair1.blend -P blender.py -p 0 0 940 1080
+	optirun blender ~/Documents/testvoronoi.blend -P blender.py -p 0 0 940 1080
 
 zhxxrun: all
 	~/Downloads/blender-2.93.3-linux-x64/blender ~/Documents/zenoStair.blend -P blender.py
